% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abcount_old}
\alias{abcount_old}
\title{Count AB (Old Version)}
\usage{
abcount_old(K, k, qj, Yj, alpha)
}
\arguments{
\item{K}{Number of Attributes as \code{unsigned int}.}

\item{k}{Present attribute level as \code{unsigned int}.}

\item{qj}{\eqn{j}th row of the \eqn{Q} matrix.}

\item{Yj}{Binary responses for individual \eqn{i} in the form of \code{vec}}

\item{alpha}{Attribute profile as a \code{mat}.}
}
\description{
What is AB? Number of attributes possessed by the individual?
}
