% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abcounts}
\alias{abcounts}
\title{Count AB}
\usage{
abcounts(N, Yj, CLASS, ETAtnokimes1ma)
}
\arguments{
\item{N}{Number of Observations given as \code{unsigned int}.}

\item{Yj}{Binary responses for individual \eqn{i} in the form of \code{vec}}

\item{CLASS}{Does the individual possess all the necessary attributes? (1 or 0)}

\item{ETAtnokimes1ma}{something?}
}
\value{
A \code{vec} containing the counts of AB parameters of the IRT
}
\description{
What is AB? Number of attributes possessed by the individual?
}
