% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alpha_matrix}
\alias{alpha_matrix}
\title{Generate the Attribute Matrix}
\usage{
alpha_matrix(Q)
}
\arguments{
\item{Q}{A \code{matrix} with dimensions \eqn{J x K}}
}
\value{
A \code{matrix} with dimensions \eqn{2^K x K}, where \eqn{2^K = C}.
}
\description{
Creates an Attribute Matrix given Q
}
\examples{
q_rand = random_Q(13, 3)
a_mat = alpha_matrix(q_rand)
}
