% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_edina.R
\name{autoplot.auto_edina}
\alias{autoplot.auto_edina}
\title{Graph the Auto EDINA Object}
\usage{
\method{autoplot}{auto_edina}(object, type = "selection", ...)
}
\arguments{
\item{object}{An \code{auto_edina} object.}

\item{type}{Kind of graph to display. Valid types: \code{"selection"}, \code{"guessing"}, or \code{"slipping"}.}
}
\description{
Presents either the fitting of model heuristics or the evolution of parameters
on a graph
}
