% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_edina.R
\name{best_model}
\alias{best_model}
\title{Extract the Best Model}
\usage{
best_model(x, ic = c("bic", "dic", "heuristic"))
}
\arguments{
\item{x}{An \code{auto_edina} object}

\item{ic}{Selection criteria}
}
\value{
An \code{edina} object
}
\description{
Extracts the best model from the \code{auto_*} search procedure.
}
