% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{bic.edina}
\alias{bic.edina}
\alias{bic}
\title{Compute the Bayesian Information Criterion (BIC)}
\usage{
\method{bic}{edina}(object, ...)

bic(object, ...)
}
\arguments{
\item{object}{An \code{edina} object}
}
\description{
Calculate BIC for EDINA models.
}
