% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bijectionvector}
\alias{bijectionvector}
\title{Bijection Vector}
\usage{
bijectionvector(K)
}
\arguments{
\item{K}{number of levels}
}
\value{
A \code{vec} with length \eqn{K}.
}
\description{
Bijection Vector
}
