% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_identifiability}
\alias{check_identifiability}
\title{Verify Q Matrix is Identifiable}
\usage{
check_identifiability(Q)
}
\arguments{
\item{Q}{The Q matrix to be checked with dimensions \eqn{K \times J}{K x J}.}
}
\value{
A \code{bool} with value either: false or true
}
\description{
Performs a check to see if Q is identifable or not.
}
