% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClassbyQmat}
\alias{ClassbyQmat}
\title{Classification Matrix by Q Matrix}
\usage{
ClassbyQmat(K)
}
\arguments{
\item{K}{Number of Attribute Levels as an \verb{unsigned integer}.}
}
\value{
A \code{mat}.
}
\description{
Construct a classification matrix by Q Matrix
}
