% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cond_threshold}
\alias{cond_threshold}
\title{Condition Threshold}
\usage{
cond_threshold(k, j, n_noks, N, K, Yj, CLASS, Q, gj, sj, ETAmatnok)
}
\arguments{
\item{k}{Present attribute level as \code{unsigned int}.}

\item{j}{Present assessment items as \code{unsigned int}.}

\item{n_noks}{Number of n okay observations?}

\item{N}{Number of Assessment Items}

\item{K}{Number of Observations}

\item{Yj}{Number of Observations}

\item{CLASS}{Does the individual possess all the necessary attributes?}

\item{gj}{The probability of guessing or the probability subject
correctly answers item \eqn{j} when at least one attribute
is lacking.}

\item{sj}{The probability of slipping or the probability of an
incorrect response for individuals with all of the required
attributes}

\item{ETAmatnok}{A variant on the \eqn{\eta} matrix.}

\item{ETA}{Alpha profile matrix.}
}
\value{
A \code{double} indicating the conditional threshold.
}
\description{
Condition Threshold
}
