% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_seconds_to_time}
\alias{convert_seconds_to_time}
\title{Convert the Number of Seconds to Time}
\usage{
convert_seconds_to_time(seconds)
}
\arguments{
\item{seconds}{A vector containing the amount of time in seconds.}
}
\value{
A vector list with entries in \code{day}, \code{hour}, \code{minute}, and \code{seconds}
}
\description{
Convert the Number of Seconds to Time
}
\details{
This function is based off of the lubridate function of a similar name.
}
