% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{dic}
\alias{dic}
\alias{dic.edina}
\title{Compute the Deviance Information Criterion (DIC)}
\usage{
dic(object, ...)

\method{dic}{edina}(object)
}
\arguments{
\item{object}{An \code{edina} object}
}
\description{
Calculate DIC for EDINA models.
}
