% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdm-package.R
\docType{package}
\name{ecdm-package}
\alias{ecdm}
\alias{ecdm-package}
\title{ecdm: Exploratory Cognitative Diagnostic Models}
\description{
Perform estimations of cognitive diagnostic models (CDMs) without
supplying a Q matrix.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

Authors:
\itemize{
  \item Steven Culpepper \email{sculpepp@illinois.edu} [copyright holder]
}

}
\keyword{internal}
