% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{edina}
\alias{edina}
\title{EDINA estimation routine}
\usage{
edina(data, k = 3, burnin = 10000, chain_length = 20000)
}
\arguments{
\item{data}{Binary responses to assessements in \code{matrix}
form with dimensions \eqn{N \times J}{N x J}.}

\item{k}{Number of Attribute Levels as a positive \code{integer}.}

\item{burnin}{Number of Observations to discard on the chain.}

\item{chain_length}{Length of the MCMC chain}
}
\value{
An \code{edina} object that contains:
}
\description{
Exploratory Determinatistic Input, Noise and Gate Model (EDINA)
}
\examples{
\dontrun{
library("tmsadata")

# Load data
data("trial_matrix", package="tmsadata")

# Coerce to matrix
trial_matrix = as.matrix(trial_matrix)

edina_model = edina(trial_matrix, k = 2)
}

}
