% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edina_Gibbs_Q}
\alias{edina_Gibbs_Q}
\title{Exploratory Determinatistic Input, Noise and Gate Model (EDINA)}
\usage{
edina_Gibbs_Q(Y, K, burnin = 1000L, chain_length = 10000L)
}
\arguments{
\item{burnin}{Number of Observations to discard on the chain.}

\item{chain_length}{Length of the MCMC chain}
}
\value{
A \code{list} containing:
\itemize{
\item \strong{GS}: Guessing
\item \strong{SS}: Slipping
\item \strong{PIs}: Latent Class Probabilities with length \eqn{K}
\item \strong{QS}: Q matrix
\item \strong{ORs}: Odds Ratio
}
}
\description{
Compute the EDINA model
}
