% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edm-package.R
\docType{package}
\name{edm-package}
\alias{edm}
\alias{edm-package}
\title{edm: Exploratory Diagnostic Models}
\description{
Perform bayesian estimations of diagnostic models (DMs)
    with (classical) or without (exploratory) supplying a Q matrix.
}
\details{
The \code{edm} package re-exports several functions from other estimation
packages developed using \code{edmcore}. As a result, the \code{edm} package is
partially a virtual package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/edm}
  \item Report bugs at \url{https://github.com/tmsalab/edm/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [copyright holder]
  \item Jeffrey A. Douglas \email{jeffdoug@illinois.edu}
}

}
\keyword{internal}
