% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errum.R
\name{errum}
\alias{errum}
\title{Exploratory reduced Reparameterized Unified Model (ErRUM)}
\usage{
errum(data, k = 3, burnin = 10000, chain_length = 10000)
}
\arguments{
\item{data}{Binary responses to assessements in \code{matrix}
form with dimensions \eqn{N \times J}{N x J}.}

\item{k}{Number of Attribute Levels as a positive \code{integer}.}

\item{burnin}{Number of Observations to discard on the chain.}

\item{chain_length}{Length of the MCMC chain}
}
\value{
An \code{errum} object that has:
\itemize{
\item \code{PISTAR}
\item \code{RSTAR}
\item \code{PIs}
\item \code{QS}
\item \code{m_Delta}
\item \code{Delta_biject}
\item \code{M2}
\item \code{M1}
\item \code{NUS}
}
}
\description{
Obtains samples from posterior distributon for the Exploratory
reduced Reparametrized Unified Model (rRUM).
}
