% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ETAmat}
\alias{ETAmat}
\title{\eqn{\eta} Matrix}
\usage{
ETAmat(K, J, Q)
}
\arguments{
\item{K}{Number of Attribute Levels}

\item{J}{Number of Assessment Items}

\item{Q}{Q Matrix with dimensions \eqn{K \times J}{K x J}.}
}
\value{
A \code{mat} with dimensions \eqn{J \times 2^K}{J x 2^K}.
}
\description{
\eqn{\eta} Matrix
}
