% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ETAmat_nok}
\alias{ETAmat_nok}
\title{\eqn{\eta} Matrix}
\usage{
ETAmat_nok(K)
}
\arguments{
\item{K}{Number of Attribute Levels}
}
\value{
A \code{cube} with dimensions \eqn{K \times 2^{K-1} \times 2^K}{J x 2^{K-1} x 2^K}.
}
\description{
\eqn{\eta} Matrix
}
