% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ETAmat_nok_one_m_ac}
\alias{ETAmat_nok_one_m_ac}
\title{\eqn{\eta} Matrix \eqn{1-\alpha_{ck}}}
\usage{
ETAmat_nok_one_m_ac(K)
}
\arguments{
\item{K}{Number of Attribute Levels}
}
\value{
A \code{cube} with dimensions \eqn{K \times 2^{K-1} \times 2^K}{J x 2^{K-1} x 2^K}.
}
\description{
\eqn{\eta} Matrix \eqn{1-\alpha_{ck}}
}
