% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{identify_check}
\alias{identify_check}
\title{Verify Q Matrix is Identifiable}
\usage{
identify_check(Q)
}
\arguments{
\item{Q}{The Q matrix to be checked with dimensions \eqn{K \times J}{K x J}.}
}
\value{
A double with value either: 0 or 1
}
\description{
Performs a check to see if Q is identifable or not.
}
