% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inv_bijectionvector}
\alias{inv_bijectionvector}
\title{Inverse Bijection Vector}
\usage{
inv_bijectionvector(K, CL)
}
\arguments{
\item{K}{number of levels}

\item{CL}{A \code{double} that controls ...}
}
\value{
A \code{vec} with length \eqn{K}.
}
\description{
Inverse Bijection Vector
}
