% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{llj}
\alias{llj}
\title{Log likelihood associated with assessment item \eqn{j}.}
\usage{
llj(N, Yj, ETAj, CLASS, gj, sj)
}
\arguments{
\item{N}{Number of Observations}

\item{Yj}{Binary responses to assessements in \code{vec} form with length \eqn{N}.}

\item{ETAj}{\eqn{\eta_j} in \code{vec} form of length \eqn{2^k}.}

\item{CLASS}{Does the individual possess all the necessary attributes?}

\item{gj}{Guessing value}

\item{sj}{Slipping Value}
}
\value{
A \code{double} containing the log likelihood
}
\description{
Log likelihood associated with assessment item \eqn{j}.
}
