% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lnlik_dina}
\alias{lnlik_dina}
\title{Compute Likelihood for DINA}
\usage{
lnlik_dina(N, J, nClass, Y, ETA, pis, gs, ss)
}
\arguments{
\item{N}{Number of Observations}

\item{J}{Number of Assessment Items}

\item{nClass}{Number of Classes typically \eqn{2^K}.}

\item{Y}{Binary responses to assessements in \code{matrix} form with
dimensions \eqn{N \times J}{N x J}.}

\item{ETA}{Q Matrix with dimensions \eqn{K x J}.}

\item{pis}{Latent Class Probabilities with length \eqn{K}}

\item{gs}{A \code{vec} describing the probability of guessing or
the probability subject correctly answers item \eqn{j} when at
least one attribute is lacking.}

\item{ss}{A \code{vec} describing the probability of slipping or
the probability of an incorrect response for individuals with
all of the required attributes}
}
\value{
The likelihood in \code{double} form.
}
\description{
Calculates the likelihood for the DINA model
}
