% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lnlik_dina_condclass}
\alias{lnlik_dina_condclass}
\title{Likelihood Function for DINA}
\usage{
lnlik_dina_condclass(N, J, Y, ETA, CLASS, pis, gs, ss)
}
\arguments{
\item{N}{Number of Observations}

\item{J}{Number of Assessment Items}

\item{Y}{Binary responses to assessements in \code{matrix} form with
dimensions \eqn{N \times J}{N x J}.}

\item{ETA}{Q Matrix with dimensions \eqn{K x J}.}

\item{CLASS}{Does the individual possess all the necessary attributes?}

\item{pis}{Latent Class Probabilities with length \eqn{K}}

\item{gs}{A \code{vec} describing the probability of guessing or
the probability subject correctly answers item \eqn{j} when at
least one attribute is lacking.}

\item{ss}{A \code{vec} describing the probability of slipping or
the probability of an incorrect response for individuals with
all of the required attributes}
}
\description{
Compute the Likelihood Function for DINA
}
