% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{model_heuristic}
\alias{model_heuristic}
\title{Model Heuristic used for Model Selection}
\usage{
model_heuristic(sample_OR, bayesian_OR)
}
\arguments{
\item{sample_OR}{A \code{matrix} with dimensions \code{J x J}
containing the empirical odds ratio.}

\item{bayesian_OR}{An \code{array} with dimensions \code{J x J}.
containing the bayesian odds ratio.}
}
\value{
A \code{double} that is the mean.
}
\description{
Computes the model heuristic for model selection.
}
