% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{new_edina}
\alias{new_edina}
\title{Construct an EDINA object}
\usage{
new_edina(x, k, burnin, chain_length, timing, sample_OR, model_fit)
}
\arguments{
\item{x}{\code{edina} object.}

\item{k}{Number of Attribute Levels as a positive \code{integer}.}

\item{burnin}{Number of Observations to discard on the chain.}

\item{chain_length}{Length of the MCMC chain}

\item{timing}{Number of Seconds that have elapsed since run time.}

\item{sample_OR}{Odds Ratio based on the trial matrix.}

\item{model_fit}{How well the model fit according to the heuristic}
}
\description{
Constructor function to satisfy the EDINA object definition
}
\keyword{internal}
