% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{new_edina_summary}
\alias{new_edina_summary}
\title{Construct a Summary EDINA object}
\usage{
new_edina_summary(m_gs, m_ss, m_pi, k, burnin, chain_length, timing, sample_OR,
  model_fit)
}
\arguments{
\item{m_gs}{Mean of the Guessing Values}

\item{m_ss}{Mean of the Slipping Values}

\item{m_pi}{Mean of the PIs}

\item{k}{Number of Attribute Levels as a positive \code{integer}.}

\item{burnin}{Number of Observations to discard on the chain.}

\item{chain_length}{Length of the MCMC chain}

\item{timing}{Number of Seconds that have elapsed since run time.}

\item{sample_OR}{Odds Ratio based on the trial matrix.}

\item{model_fit}{How well the model fit according to the heuristic}
}
\description{
Constructor function to satisfy the Summary EDINA object definition
}
\keyword{internal}
