% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OddsRatio}
\alias{OddsRatio}
\title{Compute the Odds Ratio}
\usage{
OddsRatio(N, J, Yt)
}
\arguments{
\item{N}{Number of Observations}

\item{J}{Number of Assessment Items}

\item{Yt}{Estimated binary responses to assessements in \code{matrix} form
with dimensions \eqn{N \times J}{N x J}.}
}
\value{
A \code{matrix}.
}
\description{
Calculates the Odds Ratio
}
