% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parm_update}
\alias{parm_update}
\title{Parameter Update Routine for rRUM Gibbs Sampler}
\usage{
parm_update(N, J, K, C, Y, Q, alpha, X, Smat, Gmat, pi, vv, delta0)
}
\arguments{
\item{N}{Number of Subjects}

\item{J}{Number of Assessment Items}

\item{K}{Number of Attributes}

\item{C}{Number of Classes (\eqn{2^K})}

\item{Y}{Matrix of binary responses to assessment items with
dimensions \eqn{N \times J}{N x J}.}

\item{Q}{Matrix of attributes required for each assessment
item with dimensions \eqn{J \times K}{J x K}.}

\item{alpha}{Matrix with dimensions \eqn{N \times K}{N x K} of 1s and 0s.}

\item{X}{Cube with dimensions \eqn{N \times J \times K}{N x J x K}.}

\item{Smat}{Slipping parameters}

\item{Gmat}{Guessing Parameters}

\item{pi}{Vector of PIs}

\item{vv}{Bijection vector of attributes}

\item{delta0}{Vector of length \eqn{2^K} that corresponds to the
prior values for betas and dirichlet distributions of the
classes.}
}
\description{
Parameter Update Routine for rRUM Gibbs Sampler
}
