% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parm_update_MH}
\alias{parm_update_MH}
\title{Update Parameters in Metropolis-Hastings Approach}
\usage{
parm_update_MH(N, J, K, C, Y, Q, alpha, pistar, rstar, pi, vv, delta0, delta,
  Amat)
}
\arguments{
\item{N}{Number of Subjects}

\item{J}{Number of Assessment Items}

\item{K}{Number of Attributes}

\item{C}{Number of Classes (\eqn{2^K})}

\item{Y}{Matrix of binary responses to assessment items with
dimensions \eqn{N \times J}{N x J}.}

\item{Q}{Matrix of attributes required for each assessment
item with dimensions \eqn{J \times K}{J x K}.}

\item{alpha}{Matrix with dimensions \eqn{N \times K}{N x K} of 1s and 0s.}

\item{pistar}{Vector of probabilities associated with whether
\eqn{Y_{ij} = 1} if subject \eqn{i} has all the required
attributes having length \eqn{J}.}

\item{rstar}{Matrix with dimensions \eqn{J \times K}{J x K}.}

\item{pi}{Vector of length \eqn{2^K} containing the
latent class probabilities that \eqn{Y_{ij} = 1}.}

\item{vv}{Bijection vector of length \eqn{k} for attributes.}

\item{delta0}{Vector of length \eqn{2^K} that corresponds to the
prior values for betas and dirichlet distributions of the
classes.}

\item{delta}{TBA?}

\item{Amat}{TBA?}
}
\description{
Helper functions that handles the process of updating the parameters
under the derving MH approach.
}
