% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parm_update_nomiss}
\alias{parm_update_nomiss}
\title{Update the Parameters}
\usage{
parm_update_nomiss(N, J, K, nClass, Y, ETA, gs, ss, CLASS, pis)
}
\arguments{
\item{N}{Number of Observations}

\item{J}{Number of Assessment Items}

\item{K}{Number of Attribute Levels}

\item{nClass}{Number of classes?}

\item{Y}{A \code{matrix} with dimensions \eqn{N \times J}{N x J} containing
the binary responses to assessements.}

\item{ETA}{Alpha profile matrix.}

\item{gs}{A \code{vec} describing the probability of guessing or
the probability subject correctly answers item \eqn{j} when at
least one attribute is lacking.}

\item{ss}{A \code{vec} describing the probability of slipping or
the probability of an incorrect response for individuals with
all of the required attributes}

\item{CLASS}{Does the individual possess all the necessary attributes? (1 or 0)}

\item{pis}{Latent Class Probabilities with length \eqn{K}}
}
\description{
Perform a parameter update
}
\details{
gs, ss, CLASS, and pis are updated under this function.
}
