% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{print.edina}
\alias{print.edina}
\title{Printing out the EDINA Object}
\usage{
\method{print}{edina}(x, whole = FALSE, ...)
}
\arguments{
\item{x}{An \code{edina} object}
}
\description{
Custom print method for computing the EDINA.
}
