% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pYjeq1}
\alias{pYjeq1}
\title{Probability for Equation 1?}
\usage{
pYjeq1(ETAbyQ, pis, nClass, sj, gj)
}
\arguments{
\item{ETAbyQ}{Column vectrom from ETAbyQ matrix.}

\item{pis}{Latent Class Probabilities with length \eqn{2^K}}

\item{nClass}{Classification number}

\item{sj}{Slipping Value}

\item{gj}{Guessing value}
}
\description{
Compute a probability
}
\details{
Not used in source
}
