% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_Q}
\alias{random_Q}
\title{Generate a Random Q Matrix}
\usage{
random_Q(J, K)
}
\arguments{
\item{J}{Number of Assessment Items as an \code{unsigned integer}.}

\item{K}{Number of Attribute Levels as an \code{unsigned integer}.}
}
\value{
A \code{mat}.
}
\description{
Randomly construct a Q Matrix given dimensions
}
