% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-formats.R
\name{read_items}
\alias{read_items}
\title{Import an Items Matrix}
\usage{
read_items(file, header = FALSE, sep = " ", skip = 0,
  drop_columns = NULL)
}
\arguments{
\item{file}{name, url, or \link{connections} of the file to be read}

\item{header}{a logical value to indicate if the data contains a
description field as the first row. Default is \code{FALSE}.}

\item{sep}{the field separator character. Values on each line of the
file are separated by this character. Default is white
space given by \code{sep = " "}.}

\item{skip}{the number of lines of the data file that should be
skipped before the data is read. Default is 0.}

\item{drop_columns}{columns of the item matrix that should be dropped. To
prevent the first and third columns specify
\code{drop_columns = c(1,3)}. By default, all columns are included.}
}
\value{
A \code{matrix} labeled with row names as \code{SubjectXX} and column names as \code{ItemYY}
}
\description{
Allows for a dichotmous items matrix to be imported with standard styling.
}
\details{
This function is designed to specifically read in dichotomous item matrices
into \emph{R}. The matrix must be structured with the appropriate separator.
}
