% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rRUM_Gibbs}
\alias{rRUM_Gibbs}
\title{rRUM Gibbs-based Estimation}
\usage{
rRUM_Gibbs(Y, Q, chain_length = 10000L)
}
\arguments{
\item{Y}{Matrix of binary responses to assessment items with
dimensions \eqn{N \times J}{N x J}.}

\item{Q}{Matrix of attributes required for each assessment
item with dimensions \eqn{J \times K}{J x K}.}

\item{chain_length}{Number of iterations}
}
\description{
Performs an estimation of the rRUM model using the Metropolis-Hastings
derivation.
}
