% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rRUM_MH}
\alias{rRUM_MH}
\title{rRUM Metropolis-Hastings-based Estimation}
\usage{
rRUM_MH(Y, Q, Amat, delta, chain_length = 10000L)
}
\arguments{
\item{Y}{Matrix of binary responses to assessment items with
dimensions \eqn{N \times J}{N x J}.}

\item{Q}{Matrix of attributes required for each assessment
item with dimensions \eqn{J \times K}{J x K}.}

\item{Amat}{TBA?}

\item{delta}{TBA?}

\item{chain_length}{Number of iterations}
}
\value{
A \code{list} containing:
\itemize{
\item \strong{PISTAR}: Estimates of the latent probability class.
\item \strong{RSTAR}: Estimates of the penalty parameter for missing a required
\eqn{k}.
\item \strong{PIs}: Latent Class Probabilities with length \eqn{K}
\item \strong{ACCEPTREJECT_PISTAR}: Vector of 1's or 0's indicating whether the
values for pistar was accepted or rejected.
\item \strong{ACCEPTREJECT_RSTAR}: Vector of 1's or 0's indicating whether the
values for rstar was accepted or rejected.
}
}
\description{
Performs an estimation of the rRUM model using the Metropolis-Hastings
derivation.
}
