% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simrRUM}
\alias{simrRUM}
\title{Simulate rRUM Data}
\usage{
simrRUM(N, J, K, Q, rstar, pistar, alpha)
}
\arguments{
\item{N}{Number of Subjects}

\item{J}{Number of Assessment Items}

\item{K}{Number of Attributes}

\item{Q}{Matrix of attributes required for each assessment
item with dimensions \eqn{J \times K}{J x K}.}

\item{rstar}{Matrix of penalities for missing an attribute on an item
with dimensions \eqn{J \times K}{J x K}.}

\item{pistar}{Vector of probabilities associated with whether
\eqn{Y_{ij} = 1} if subject \eqn{i} has all the required
attributes having length \eqn{J}.}

\item{alpha}{Matrix with dimensions \eqn{N \times K}{N x K} of 1s and 0s.}
}
\description{
Generate data that falls underneath a rRUM modeling paradigm.
}
