\name{startEDM}
\alias{startEDM}
\title{Launch EDM User Interface}
\usage{
startEDM()
}
\value{
Nothing
}
\description{
startEDM() loads interactive user interface built using R shiny.
}
\details{
Start Exame Database Manager

blabla
}
\examples{
\dontrun{
library(shiny)
startEDM()
}
}
\keyword{startEDM}

