% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{summary.edina}
\alias{summary.edina}
\title{Summarizing the EDINA Object}
\usage{
\method{summary}{edina}(object, ...)
}
\arguments{
\item{object}{An \code{edina} object}
}
\description{
Summarizing the EDINA Object
}
