% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateQ_DINA}
\alias{updateQ_DINA}
\title{Update the Q}
\usage{
updateQ_DINA(Q, Y, alpha, ss, gs)
}
\arguments{
\item{Q}{Q Matrix with dimensions \eqn{K x J}.}

\item{Y}{Binary responses to assessements in \code{matrix} form with
dimensions \eqn{N \times J}{N x J}.}

\item{alpha}{Profile Matrix}

\item{ss}{A \code{vec} describing the probability of slipping or
the probability of an incorrect response for individuals with
all of the required attributes}

\item{gs}{A \code{vec} describing the probability of guessing or
the probability subject correctly answers item \eqn{j} when at
least one attribute is lacking.}
}
\description{
Updation step for DINA
}
