% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateQ_DINA_new}
\alias{updateQ_DINA_new}
\title{Simulate Binary Responses for DINA Model}
\usage{
updateQ_DINA_new(N, K, J, Q, Y, CLASS, ss, gs, vj, ETAmatnok, a_by_q, vv)
}
\arguments{
\item{N}{Number of Observations}

\item{K}{Number of Attribute Levels as an \code{unsigned integer}.}

\item{J}{Number of Assessment Items}

\item{Q}{Q Matrix with dimensions \eqn{K x J}.}

\item{Y}{Binary responses to assessements in \code{matrix} form with
dimensions \eqn{N \times J}{N x J}.}

\item{CLASS}{Does the individual possess all the necessary attributes?}

\item{ss}{A \code{vec} describing the probability of slipping or
the probability of an incorrect response for individuals
with all of the required attributes}

\item{gs}{A \code{vec} describing the probability of guessing or
the probability subject correctly answers item \eqn{j} when
at least one attribute is lacking.}

\item{vj}{A \code{vec} containing the powers of 2.}

\item{ETAmatnok}{A variant on the \eqn{\eta} Matrix.}

\item{a_by_q}{Classification Matrix by a \eqn{Q} Matrix.}

\item{vv}{Bijection vector with respect to \eqn{K}.}
}
\description{
Calculates the Odds Ratio
}
\details{
No return is done here as the update is done by reference.
}
