% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdm-package.R
\docType{package}
\name{ecdm-package}
\alias{ecdm}
\alias{ecdm-package}
\title{ecdm: Exploratory Cognitative Diagnostic Models}
\description{
Perform bayesian estimations of cognitive diagnostic models (CDMs)
    with (classical) or without (exploratory) supplying a Q matrix.
}
\details{
The \code{ecdm} package re-exports several functions from other estimation
packages developed by \code{ecdmcore}. As a result, the \code{ecdm} package is
partially a virtual package.
}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} (0000-0003-2826-8458) [copyright holder]

Authors:
\itemize{
  \item Steven Culpepper \email{sculpepp@illinois.edu} [copyright holder]
  \item Jeffrey Douglas \email{jeffdoug@illinois.edu} [copyright holder]
}

}
