% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edina.R
\name{print.summary_edina}
\alias{print.summary_edina}
\title{Printing out the Summary EDINA Object}
\usage{
\method{print}{summary_edina}(x, whole = FALSE, ...)
}
\arguments{
\item{x}{A \code{summary_edina} object}
}
\description{
Printing out the Summary EDINA Object
}
