% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-exports.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sim_dina}
\alias{reexports}
\alias{sim_rrum}
\alias{reexports}
\alias{dina}
\alias{reexports}
\alias{rrum}
\alias{reexports}
\alias{autoplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dina}{\code{\link[dina]{dina}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{rrum}{\code{\link[rrum]{rrum}}}

  \item{simcdm}{\code{\link[simcdm]{sim_dina}}, \code{\link[simcdm]{sim_rrum}}}
}}

